Restructure: move some helpers from the SurveyWidget to the Survey feature
